refatora estrutura de diretorios
