--- Processo/src/main/java/com/mycompany/processo/Cpu.java
@@ -1,10 +1,0 @@
-
-package com.mycompany.processo;
-
-public class Cpu {
-    public Processo processoAtual;
-
-    public Cpu(Processo processoAtual) {
-        this.processoAtual = processoAtual;
-    }
-}
