--- Processo/src/main/java/com/mycompany/processo/Simulador.java
@@ -1,11 +1,0 @@
-
-package com.mycompany.processo;
-
-import java.util.List;
-
-public class Simulador {
-    public List<Cpu> cpus;
-    public Fila<Processo> filaReal;
-    public List<Fila<Processo>> filasUsuario;
-    public List<Processo> processosPendentes;
-}
