+++ src/main/java/componentes/Cpu.java
@@ -1,0 +1,10 @@
+
+package componentes;
+
+public class Cpu {
+    public Processo processoAtual;
+
+    public Cpu(Processo processoAtual) {
+        this.processoAtual = processoAtual;
+    }
+}
