+++ src/main/java/componentes/Processo.java
@@ -1,4 +1,4 @@
-package com.mycompany.processo;
+package componentes;
 import java.io.BufferedReader;
 import java.io.FileReader;
 import java.io.IOException;
