+++ src/main/java/componentes/Simulador.java
@@ -1,0 +1,11 @@
+
+package componentes;
+
+import java.util.List;
+
+public class Simulador {
+    public List<Cpu> cpus;
+    public Fila<Processo> filaReal;
+    public List<Fila<Processo>> filasUsuario;
+    public List<Processo> processosPendentes;
+}
