+++ src/main/java/main/Main.java
@@ -1,4 +1,4 @@
-package org.example;
+package main;
 
 import relatorio.Relatorio;
